WIP: refactoring Mojo to re-use base S3 upload functionality
