+++ src/main/java/com/edmunds/tools/databricks/maven/UploadToS3Mojo.java
@@ -37,9 +37,8 @@ import org.apache.maven.plugins.annotations.Parameter;
  * Uploads an artifact to s3.
  */
 @Mojo(name = "upload-to-s3", defaultPhase = LifecyclePhase.DEPLOY)
-public class UploadToS3Mojo extends BaseDatabricksMojo {
+public class UploadToS3Mojo extends BaseDatabricksS3Mojo {
 
-    protected AmazonS3 s3Client;
     /**
      * The local file to upload.
      */
@@ -47,75 +46,8 @@ public class UploadToS3Mojo extends BaseDatabricksMojo {
         defaultValue = "${project.build.directory}/${project.build.finalName}.${project.packaging}")
     protected File file;
 
-    protected String createSourceFilePath() throws MojoExecutionException {
-        return createDeployedArtifactPath();
-    }
-
     @Override
-    public void execute() throws MojoExecutionException {
-        if (file.exists()) {
-            AmazonS3URI uri = new AmazonS3URI(createSourceFilePath());
-            String bucket = uri.getBucket();
-            String key = uri.getKey();
-            try {
-                PutObjectRequest putObjectRequest =
-                    new PutObjectRequest(bucket, key, file)
-                        .withCannedAcl(CannedAccessControlList.BucketOwnerFullControl);
-                putObjectRequest.setGeneralProgressListener(new LoggingProgressListener(getLog(), file.length()));
-
-                getLog().info(String.format("Starting upload for bucket: [%s] key: [%s], file: [%s]",
-                    bucket, key, file.getPath()));
-
-                getS3Client().putObject(putObjectRequest);
-            } catch (SdkClientException e) {
-                throw new MojoExecutionException(String.format("Could not upload file: [%s] to bucket: [%s] with "
-                    + "remote prefix: [%s]", file.getPath(), bucket, key), e);
-            }
-        } else {
-            getLog().warn(String.format("Target upload file does not exist, skipping: [%s]", file.getPath()));
-        }
+    protected File getSourceFile() {
+        return file;
     }
-
-    protected AmazonS3 getS3Client() {
-        if (s3Client == null) {
-            AWSCredentialsProvider credentialsProvider = DefaultAWSCredentialsProviderChain.getInstance();
-            s3Client = AmazonS3ClientBuilder
-                .standard()
-                .withRegion(databricksRepoRegion)
-                .withCredentials(credentialsProvider)
-                .build();
-
-        }
-        return s3Client;
-    }
-
-    private static class LoggingProgressListener implements ProgressListener {
-
-        private final Log log;
-        private final long size;
-        private double progress = 0;
-
-        private long startTime = System.currentTimeMillis();
-        private long lastTimeLogged = System.currentTimeMillis();
-
-        public LoggingProgressListener(Log log, long size) {
-            this.log = log;
-            this.size = size;
-        }
-
-        @Override
-        public void progressChanged(ProgressEvent progressEvent) {
-            progress += progressEvent.getBytesTransferred();
-
-            double percent = (progress / size) * 100;
-            long now = System.currentTimeMillis();
-            if (now - lastTimeLogged > 2000) {
-                log.info(
-                    String.format("Transferred %.2f%% to s3, total run time: %ss.", percent, (now - startTime) / 1000));
-                lastTimeLogged = now;
-            }
-
-        }
-    }
-
 }
